refacto to move queries in view instead of components
