--- lol-esport-stats/src/components/TopBar.svelte
@@ -1,3 +1,0 @@
-<div>
-    <p>La barre de fou</p>
-</div>
